requirements.txt
beautifulsoup4                                  
bs4                                                           
pip                                              
pycparser                                     
requests                                                   
ScraperFC                                           
standard-imghdr                                   
tweepy                
python-dotenv
LanusStats